% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{ABC}
\alias{ABC}
\title{Perform an ABC analysis}
\usage{
ABC(nsims, prior_function, sim_function, sumstat_functions, obs_data,
  dist = "weightedEuclidean", nacc = 200, output = "accepted",
  nCores = parallel::detectCores())
}
\arguments{
\item{nsims}{How many simulations to perform.}

\item{prior_function}{A function with no arguments returing a vector of parameters drawn from the prior. Ideally this vector should be named.}

\item{sim_function}{A function taking a vector of parameters as an argument and returning a model data object.}

\item{sumstat_functions}{A list of functions. Each should take a model data object as argument and return a summary statistic (or a vector, see later), which ideally is named. Sometimes it is convenient to use the same calculations for several summaries. In this case a function can return a vector of summaries. Again these should ideally be named.}

\item{obs_data}{Observed model data object.}

\item{dist}{What distance function to use. Currently choose either \code{"weightedEuclidean"} or \code{"rank"}.}

\item{nacc}{How many simulations to accept.}

\item{output}{Specify what output to return. Currently choose either \code{"accepted"} or \code{"all"}.}

\item{nCores}{Number of parallel cores to use.}
}
\value{
A data frame containing parameters and summaries. If \code{output=="accepted"} then only the accepted simulations are returned. Otherwise all simulations are returns and an \code{output=="accepted"} column shows if they were accepted.
}
\description{
Perform an ABC analysis, including simulation of data
}

