% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{cor_overall}
\alias{cor_overall}
\title{Rainfall autocorrelation}
\usage{
cor_overall(x)
}
\arguments{
\item{x}{Numeric vector of rainfall totals}
}
\value{
Lag 1 autocorrelation
}
\description{
Rainfall autocorrelation
}

