% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{cov_overall}
\alias{cov_overall}
\title{Rainfall autocovariance}
\usage{
cov_overall(x)
}
\arguments{
\item{x}{Numeric vector of rainfall totals}
}
\value{
Lag 1 autocovariance
}
\description{
Rainfall autocovariance
}

