% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{mean_overall}
\alias{mean_overall}
\title{Overall mean rainfall}
\usage{
mean_overall(x)
}
\arguments{
\item{x}{Numeric vector of rainfall totals}
}
\value{
Mean rainfall
}
\description{
Overall mean rainfall
}

