% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{mean_wet}
\alias{mean_wet}
\title{Mean rainfall on wet days}
\usage{
mean_wet(x)
}
\arguments{
\item{x}{Numeric vector of rainfall totals}
}
\value{
Mean of rainfall on days with non-zero rain
}
\description{
Mean rainfall on wet days
}

