% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{prop_dry}
\alias{prop_dry}
\title{Proportion of dry days}
\usage{
prop_dry(x)
}
\arguments{
\item{x}{Numeric vector of rainfall totals}
}
\value{
Proportion of days with zero rainfall
}
\description{
Proportion of dry days
}

