% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{rejABC}
\alias{rejABC}
\title{Perform an ABC analysis given simulations}
\usage{
rejABC(stats, obs, subset = 1:ncol(stats), dist = "weightedEuclidean",
  nacc = 200, nCores = detectCores())
}
\arguments{
\item{stats}{A matrix of simulated summary statistics.}

\item{obs}{A vector of observed summary statistics.}

\item{subset}{A numeric vector specifying which summaries should be used.}

\item{dist}{What distance function to use. Currently choose either \code{"weightedEuclidean"} or \code{"rank"}.}

\item{nacc}{How many simulations to accept.}

\item{nCores}{Number of parallel cores to use.}
}
\value{
A vector specifying which rows of \code{stats} were accepted.
}
\description{
Perform an ABC analysis given simulations
}

