% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{rejABC_1stat}
\alias{rejABC_1stat}
\title{Perform an ABC analysis given one simulated summary}
\usage{
rejABC_1stat(stats, obs, dist = "weightedEuclidean", nacc = 200)
}
\arguments{
\item{stats}{A vector of simulated summary statistics.}

\item{obs}{The observed summary statistic.}

\item{dist}{What distance function to use. Currently choose either "weightedEuclidean" or "rank".}

\item{nacc}{How many simulations to accept.}
}
\value{
A vector specifying which elements of \code{stats} were accepted.
}
\description{
Perform an ABC analysis given simulations of a single summary statistic
}

