% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteredPoissonProcess.R
\name{sim_simple_data}
\alias{sim_simple_data}
\title{Simulate a rainfall time series from a simplified model}
\usage{
sim_simple_data(lambda, mu_x, eta, dur, interval, bartlett_lewis = TRUE)
}
\arguments{
\item{lambda}{Rate of storm arrival.}

\item{mu_x}{Mean rain cell depth}

\item{eta}{1/mean rain cell duration}

\item{dur}{Number of time units over which to simulate}

\item{interval}{How many time units to aggregate into a single
observation}

\item{bartlett_lewis}{If \code{TRUE} (the default) simulate from
a Bartlett-Lewis process. Otherwise simulate from a Neymann-Scott
process}
}
\description{
Simulate a rainfall time series from a simplified model
}

