% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{simABC}
\alias{simABC}
\title{Simulate data to be used in ABC}
\usage{
simABC(nsims, prior_function, sim_function, sumstat_functions,
  nCores = detectCores())
}
\arguments{
\item{nsims}{How many simulations to perform.}

\item{prior_function}{A function with no arguments returing a vector of parameters drawn from the prior. Ideally this vector should be named.}

\item{sim_function}{A function taking a vector of parameters as an argument and returning a model data object.}

\item{sumstat_functions}{A list of functions. Each should take a model data object as argument and return a summary statistic (or a vector, see later), which ideally is named. Sometimes it is convenient to use the same calculations for several summaries. In this case a function can return a vector of summaries. Again these should ideally be named.}

\item{nCores}{Number of parallel cores to use.}
}
\value{
A matrix whose rows are separate simulations. Each row contains the simulations parameters followed by the summaries.
}
\description{
Parameters are drawn from the prior distribution, fed in to the model to produce data, which is processed to produce summary statistics. This process is iterated.
}

