% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{var_overall}
\alias{var_overall}
\title{Overall rainfall variance}
\usage{
var_overall(x, ...)
}
\arguments{
\item{x}{Numeric vector of rainfall totals}
}
\value{
Rainfall variance
}
\description{
Overall rainfall variance
}

