% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{var_wet}
\alias{var_wet}
\title{Variance of rainfall on wet days}
\usage{
var_wet(x, cutoff = 0)
}
\arguments{
\item{x}{Numeric vector of rainfall totals}

\item{cutoff}{If cutoff>0, rain day defined as day where rain>=cutoff, otherwise rain day defined as day where rain>0}
}
\value{
Variance of rainfall on rain days
}
\description{
Variance of rainfall on wet days
}

