% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{prop_dry}
\alias{prop_dry}
\title{Proportion of dry days}
\usage{
prop_dry(x, cutoff = 0)
}
\arguments{
\item{x}{Numeric vector of rainfall totals}

\item{cutoff}{If cutoff>0, dry day defined as day where rain<cutoff, otherwise dry day defined as day where rain=0}
}
\value{
Proportion of dry days
}
\description{
Proportion of dry days
}

