% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{streaks}
\alias{streaks}
\title{Summaries related to streaks of consecutive rainy days}
\usage{
streaks(x)
}
\arguments{
\item{x}{Numeric vector of rainfall totals}
}
\value{
A vector of \code{mean_dur}, mean streak duration; \code{sd_dur}, standard deviation of streak durations; \code{mean_start_gap}, mean times between start times.
Where not meaningfully defined the default values of these are 0, 0 and \code{length(x) + 1}.
}
\description{
Summaries related to streaks of consecutive rainy days
}

